requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.37
asyncpg==0.30.0
alembic==1.14.1
pydantic-settings==2.7.1
httpx==0.28.1
PyJWT==2.9.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
python-multipart==0.0.20
slowapi==0.1.9
